Windows_Paths_Converted test has now been removed
This test will always fail on OSes which use forward slashes as path separators, because on those systems, NormalisePath tries to replace *forward* slashes in the path passed to IsMatch (there aren't any, so the backslashes remain and the test fails).

In hindsight this didn't really serve any useful purpose anyway... so removing it was the best option.